Rock Crushing
It is now possible to tap the rock when interacting with it and break it, revealing one of three "artifacts" with a name and description which is shown. The rock shakes when tapped and plays a short particle effect which increases with each tap. Idea is to define some relevant artifacts related to signs of life on mars and make a kind of glossary of the things you have found yet which the player can read previously found object descriptions.