Rover Extra Polish
There is now sound effects for most actions. The text is written in sequence when analysing. Plane texture in Mars. Broken rock model when crushing. More particles. Hints for question screen. Worked on making the questions work with the questionmanager script but things changed so I am waiting to make it work till we have talked about the new features and how they work.