Updated Model Target Menu
Tested different tracking optimizations and made the mesh overlaying the model target have a holographic effect. Solar system pauses when selecting a planet using a solarsystem manager. Can't get collision to work on the model target, because as soon as any mesh is loaded it begins to lag and run insanely badly on the phone. Have tried using the water tracking script but still crashes. Currently it does nothing when colliding the target into planets, which needs to be fixed.