Quality of Life Rocket Scene
The relative speed of planets can be changed in the solar system manager and is set to a quarter of the one before. Mars and Earth is twice as big, while the rocket is half the size. There is now a grace period after closing a UI panel for opening the same one again. + several small fixes.