Solar System + Rover Finishes
Most of the final touches for the solar system menu and the rover scene.

Logic to keep people from selecting the same planet twice + red hologram effect. Tutorial text for both scenes that tell the user what to do the first time. UI moved and polished to fit the current aesthetic

+ other small changes and fixes